pipeline: Split the scheduler into a separate file
